requirements.txt
guidebook-api-python==0.0.2
requests==2.11.1